engine/pyproject.toml
[project]
name = "engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.0",
    "py-redis>=1.1.1",
    "pydantic>=2.12.5",
    "uvicorn>=0.41.0",
]
